updated structure with custom hooks and improved naming
